``` refactor(block):重构方块相关初始化包结构将 ModBlocks、ModBlockEntities 等类移动到对应的 block 子包中，以优化代码组织结构并提高模块清晰度。
refactor(fluid): 重构流体相关初始化包结构

将 ModFluids 类移动至 fluid 子包下，并更新相应导入路径。

refactor(item): 重构物品相关初始化包结构

将 ModItems 和 ModItemGroups 类移动至 item 子包下，
同时调整其引用路径以匹配新的包结构。

refactor(util):重命名工具类及迁移至 util 包

将 GetEnchantments 类更名为 EnchantmentUtil 并移至 util 包内，
统一管理附魔相关的工具方法。

chore(deps): 移除 cloth-config依赖并升级 anvilcraft 版本从 dependencies.gradle 中删除 clothConfig 的引入；
在 libs.versions.toml 中将 anvilCraft 升级为 pre-release.18 版本。```
