+++ src/main/java/me/theabab2333/headtap/HeadTap.java
@@ -3,12 +3,12 @@ package me.theabab2333.headtap;
 import com.tterrag.registrate.Registrate;
 import me.theabab2333.headtap.api.tooltip.ItemTooltipManager;
 import me.theabab2333.headtap.data.HeadTapDataGen;
-import me.theabab2333.headtap.init.ModBlockEntities;
-import me.theabab2333.headtap.init.ModBlocks;
+import me.theabab2333.headtap.init.block.ModBlockEntities;
+import me.theabab2333.headtap.init.block.ModBlocks;
 import me.theabab2333.headtap.init.ModEntities;
-import me.theabab2333.headtap.init.ModFluids;
-import me.theabab2333.headtap.init.ModItemGroups;
-import me.theabab2333.headtap.init.ModItems;
+import me.theabab2333.headtap.init.fluid.ModFluids;
+import me.theabab2333.headtap.init.item.ModItemGroups;
+import me.theabab2333.headtap.init.item.ModItems;
 import me.theabab2333.headtap.init.ModMenuTypes;
 import me.theabab2333.headtap.init.ModNetworks;
 import me.theabab2333.headtap.init.ModRecipeTypes;
