+++ src/main/java/me/theabab2333/headtap/block/entity/BuilderBlockEntity.java
@@ -10,8 +10,8 @@ import dev.dubhe.anvilcraft.recipe.multiblock.BlockPattern;
 import dev.dubhe.anvilcraft.recipe.multiblock.MultiblockRecipe;
 import lombok.Getter;
 import me.theabab2333.headtap.block.BuilderBlock;
-import me.theabab2333.headtap.init.ModBlockEntities;
-import me.theabab2333.headtap.init.ModBlocks;
+import me.theabab2333.headtap.init.block.ModBlockEntities;
+import me.theabab2333.headtap.init.block.ModBlocks;
 import me.theabab2333.headtap.init.ModMenuTypes;
 import me.theabab2333.headtap.inventory.BuilderMenu;
 import net.minecraft.MethodsReturnNonnullByDefault;
