+++ src/main/java/me/theabab2333/headtap/block/entity/ResinExtractorBlockEntity.java
@@ -1,8 +1,8 @@
 package me.theabab2333.headtap.block.entity;
 
-import me.theabab2333.headtap.init.ModBlockEntities;
-import me.theabab2333.headtap.init.ModBlocks;
-import me.theabab2333.headtap.init.ModFluids;
+import me.theabab2333.headtap.init.block.ModBlockEntities;
+import me.theabab2333.headtap.init.block.ModBlocks;
+import me.theabab2333.headtap.init.fluid.ModFluids;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
