+++ src/main/java/me/theabab2333/headtap/data/tags/ItemTagLoader.java
@@ -2,7 +2,7 @@ package me.theabab2333.headtap.data.tags;
 
 import com.tterrag.registrate.providers.RegistrateTagsProvider;
 import dev.dubhe.anvilcraft.init.item.ModItemTags;
-import me.theabab2333.headtap.init.ModItems;
+import me.theabab2333.headtap.init.item.ModItems;
 import net.minecraft.world.item.Item;
 import org.jetbrains.annotations.NotNull;
